Update: Deleted StatusCodeGenerator.java, StatusCodeGeneratorImpl.java and restructured
