+++ commons/src/main/java/com/twozo/commons/exception/service/StatusCode.java
@@ -1,4 +1,4 @@
-package com.twozo.commons.exception.status.code.service;
+package com.twozo.commons.exception.service;
 
 /**
  * <p>
